틀린문제까지 검산 다 완료 저장까지 완료

+++ src/main/java/com/moaguide/domain/quiz/QuestionRepository.java
@@ -14,22 +14,22 @@ import java.util.List;
 public interface QuestionRepository extends JpaRepository<Question, Long> {
 
     @Query("SELECT new com.moaguide.dto.NewDto.customDto.QuestionDto(q.explanation,q.choice1,q.choice2,q.choice3,q.choice4,q.choice5) FROM Question q where q.quizId = :id  ORDER BY CASE q.type WHEN 'b' THEN 1 WHEN 'c' THEN 2 WHEN 'a' THEN 3 END")
-    List<QuestionDto> findByTypeB(@Param("id") Integer id);
+    List<QuestionDto> findByTypeB(@Param("id") long id);
 
     @Query("SELECT new com.moaguide.dto.NewDto.customDto.QuestionDto(q.explanation,q.choice1,q.choice2,q.choice3,q.choice4,q.choice5) FROM Question q where q.quizId = :id  ORDER BY CASE q.type WHEN 'c' THEN 1 WHEN 'a' THEN 2 WHEN 'b' THEN 3 END")
-    List<QuestionDto> findByTypeC(@Param("id") Integer id);
+    List<QuestionDto> findByTypeC(@Param("id") long id);
 
     @Query("SELECT new com.moaguide.dto.NewDto.customDto.QuestionDto(q.explanation,q.choice1,q.choice2,q.choice3,q.choice4,q.choice5) FROM Question q where q.quizId = :id  ORDER BY CASE q.type WHEN 'a' THEN 1 WHEN 'b' THEN 2 WHEN 'c' THEN 3 END")
-    List<QuestionDto> findByTypeA(@Param("id") Integer id);
+    List<QuestionDto> findByTypeA(@Param("id") long id);
 
     @Query("SELECT new com.moaguide.dto.NewDto.customDto.QuestionCheckResponseDto(q.id,q.solution,q.score) FROM Question q where q.quizId = :id  ORDER BY CASE q.type WHEN 'b' THEN 1 WHEN 'c' THEN 2 WHEN 'a' THEN 3 END")
-    List<QuestionCheckResponseDto> findByTypeBcheck(Integer id);
+    List<QuestionCheckResponseDto> findByTypeBcheck(long id);
 
     @Query("SELECT new com.moaguide.dto.NewDto.customDto.QuestionCheckResponseDto(q.id,q.solution,q.score) FROM Question q where q.quizId = :id  ORDER BY CASE q.type WHEN 'c' THEN 1 WHEN 'a' THEN 2 WHEN 'b' THEN 3 END")
-    List<QuestionCheckResponseDto> findByTypeCcheck(Integer id);
+    List<QuestionCheckResponseDto> findByTypeCcheck(long id);
 
     @Query("SELECT new com.moaguide.dto.NewDto.customDto.QuestionCheckResponseDto(q.id,q.solution,q.score) FROM Question q where q.quizId = :id  ORDER BY CASE q.type WHEN 'a' THEN 1 WHEN 'b' THEN 2 WHEN 'c' THEN 3 END")
-    List<QuestionCheckResponseDto> findByTypeAcheck(Integer id);
+    List<QuestionCheckResponseDto> findByTypeAcheck(long id);
 
     @Query("SELECT new com.moaguide.dto.NewDto.customDto.QuestionLinkDto(q.explanation,q.choice1,q.choice2,q.choice3,q.choice4,q.choice5,q.link) FROM Question q where q.id in :fail ")
     List<QuestionLinkDto> findByFailId(@Param("fail")List<Long> faillist);
